Adding Editing or Deleting truck is done

+++ truck-fleet/src/stores/Trucks/TrucksStore.tsx
@@ -18,18 +18,26 @@ import { db } from "@/lib/Firebase";
 
 interface TruckStore {
 	trucks: Truck[];
+	selectedTruck: null | Truck;
+	isDeletingTruck: boolean;
+	isEditingTruck: boolean;
 	loadTrucks: (companyId: string) => () => void; // returns unsubscribe function
-	editTruck: (truckId: string, data: Partial<Truck>) => Promise<void>;
-	deleteTruck: (truckId: string) => Promise<void>;
+	editTruck: (confirmed: boolean, data: Partial<Truck>) => Promise<void>;
+	deleteTruck: (confirmed: boolean) => Promise<void>;
 	createTruck: (
 		companyId: string,
 		data: Omit<Truck, "id" | "createdAt">,
 	) => Promise<void>;
+	deleteTruckDialog: (truckId: string) => void;
+	editTruckDialog: (truckId: string) => void;
 }
 
 export const useTruckStore = create<TruckStore>()(
-	devtools((set) => ({
+	devtools((set, get) => ({
 		trucks: [],
+		isDeletingTruck: false,
+		isEditingTruck: false,
+		selectedTruck: null,
 		loadTrucks: (companyId: string) => {
 			const trucksRef = collection(db, "trucks");
 			const trucksQuery = query(
@@ -45,14 +53,39 @@ export const useTruckStore = create<TruckStore>()(
 			});
 			return unsubscribe;
 		},
-		editTruck: async (truckId: string, data: Partial<Truck>) => {
-			const truckDocRef = doc(db, "trucks", truckId);
+		editTruckDialog: async (truckId: string) => {
+			const selectedTruck = get().trucks.find((truck) => truck.id === truckId);
+
+			set({ isEditingTruck: true, selectedTruck });
+		},
+		editTruck: async (confirmed: boolean, data: Partial<Truck>) => {
+			const selectedTruck = get().selectedTruck;
+			if (selectedTruck === null || !confirmed) {
+				set({ isEditingTruck: false, selectedTruck: null });
+				return;
+			}
+
+			const truckDocRef = doc(db, "trucks", selectedTruck.id);
 			await updateDoc(truckDocRef, data);
 		},
-		deleteTruck: async (truckId: string) => {
-			const truckDocRef = doc(db, "trucks", truckId);
-			await deleteDoc(truckDocRef);
+		deleteTruckDialog: async (truckId: string) => {
+			const selectedTruck = get().trucks.find((truck) => truck.id === truckId);
+
+			set({ isDeletingTruck: true, selectedTruck });
 		},
+		deleteTruck: async (confirmed: boolean) => {
+			const selectedTruck = get().selectedTruck;
+
+			if (selectedTruck === null || !confirmed) {
+				set({ isDeletingTruck: false, selectedTruck: null });
+				return;
+			}
+
+			const truckDocRef = doc(db, "trucks", selectedTruck.id);
+			await deleteDoc(truckDocRef);
+			set({ isDeletingTruck: false, selectedTruck: null });
+		},
+
 		createTruck: async (
 			companyId: string,
 			data: Omit<Truck, "id" | "createdAt">,
